+++ both/views/pages/PageReports.jsx
@@ -8,8 +8,10 @@ PageReports = React.createClass({
     <div className="wrapper wrapper-content animated fadeIn">
         <div className="row">
             <div className="col-lg-12">
-	            <Box title="Large Reports Selction Grid">
-	            	<div style={{height:"800px"}}></div>
+	            <Box title="Reports">
+	            	<div style={{height:"800px"}}>
+	            		<img src="img/reports-screen.png"/>
+	            	</div>
 	            </Box>
 			</div>
 		</div>